third_party/blink/web_tests/wpt_internal/environment-integrity/android-specific.sub.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type="module">
  "use strict";
  import { getTokenFromOtherOrigin } from "./resources/crossOriginUtils.js";

  // TODO(bewise): We should have a test looking at the actual CBOR token contents when we
  // get a response from Android.
  // We can validate this against the actual attester.

  promise_test(async () => {
    const selfToken = await navigator.getEnvironmentIntegrity("same binding");
    const otherToken = await navigator.getEnvironmentIntegrity("same binding");

    // We will likely need to make this test more specific when we have actual token
    // responses.
    assert_equals(selfToken.encode(), otherToken.encode());
  }, "getEnvironmentIntegrity should return same token for same origin with same content binding");

  promise_test(async () => {
    const selfToken = await navigator.getEnvironmentIntegrity("first binding");
    const otherToken = await navigator.getEnvironmentIntegrity("different binding");

    assert_not_equals(selfToken.encode(), otherToken);
  }, "getEnvironmentIntegrity should return different token for same origin with different content bindings");

  promise_test(async () => {
    const selfToken = await navigator.getEnvironmentIntegrity("same binding");
    const otherToken = await getTokenFromOtherOrigin("https://{{hosts[alt][]}}:{{ports[https][0]}}", "same binding");

    assert_not_equals(selfToken.encode(), otherToken);
  }, "getEnvironmentIntegrity should always return different token between different origins");

</script>
